fix empty source

+++ functions/Invoke-ChocoBuildInstall.ps1
@@ -23,7 +23,9 @@ See the %install section below for more.
 
   # Deploy the content to root path
   $null = New-Item -Force -ItemType Directory $PackageBuildRootFilesPath
-  $null = Copy-Item -Force -Recurse "${PackageBuildPath}\*" "${PackageBuildRootFilesPath}"
+  if (Test-Path $PackageBuildPath) {
+    $null = Copy-Item -Force -Recurse "${PackageBuildPath}\*" "${PackageBuildRootFilesPath}"
+  }
 
 }
